refactor(record): rename substitution files to follow resource-based naming convention
- Rename create-substitution.usecase.ts to substitution.usecase.ts
- Rename create-substitution.controller.ts to substitution.controller.ts
- Update import paths in API route file and inversify.config.ts
- Maintain all existing functionality while following consistent naming pattern

This change aligns substitution component naming with other record modules
that use resource-based naming (set.usecase.ts, rally.usecase.ts) rather than
action-based naming (create-*.usecase.ts).
