+++ src/applications/usecases/record/substitution.usecase.ts
